requirements.txt
llama-cpp-python
chromadb
sentence-transformers

